requirements.txt
# Core FastAPI dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Database - PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Security & Authentication
cryptography==43.0.3
bcrypt==4.2.1
python-jose[cryptography]==3.3.0

# Request handling
python-multipart==0.0.12

# Configuration
python-dotenv==1.0.1

# Data validation
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP requests
requests==2.32.3

# QR Code generation
qrcode[pil]==8.0
Pillow==11.0.0

# Push Notifications
pywebpush==2.1.2

# WebSockets
python-socketio==5.15.0